Fetching currency rates from the backend is now working.
I had a weird bug that inflicted error every time the app was loading. There was a function that was filtering only the needed currency infos out of hudreds. Even though it was called only when it had something to process, only an empty array was passed to CurrencyTable component.

The workaround for this was to turn that function into a variable, so it will be assigned already at the render.